requirements.txt
amqp==5.0.6
asgiref==3.4.1        
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.12
django-extensions==3.1.5
djangorestframework==3.12.4
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
kombu==5.2.1
loguru==0.5.3
lxml==4.7.1
numpy==1.21.4
pandas==1.3.5
prompt-toolkit==3.0.22
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
scikit-learn==1.0.2
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.1
threadpoolctl==3.1.0
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
win32-setctime==1.1.0
xgboost==1.5.2